refactor(wren-ui): make code block to factory pattern
