+++ wren-ui/src/components/editor/AceEditor.tsx
@@ -1,6 +1,7 @@
 import AceEditor from 'react-ace';
 
 import 'ace-builds/src-noconflict/mode-sql';
+import 'ace-builds/src-noconflict/mode-json';
 import 'ace-builds/src-noconflict/theme-tomorrow';
 import 'ace-builds/src-noconflict/ext-language_tools';
